requirements.txt
pandas>=1.5.0
python-dateutil>=2.8.0
pytest>=7.0.0
pytest-cov>=4.0.0
black>=23.0.0
flake8>=6.0.0